+++ src/proposal.jl
@@ -3,7 +3,7 @@ export getProposalAddress, paramProposal
 """
 Optimizes `paramProposal` by providing compact way to access trace address symbols
 """
-function getProposalAddress(name::String; i::Int = -1, j::Int = -1)
+function getProposalAddress(name::String; i::Int64=-1, j::Int64=-1)
     proposalAddresses = Dict(
         "uNoise" => :uNoise,
         "tNoise" => :tNoise,
